Use RawConfigParser instead of SafeConfigParser
This change keeps get_credentials() from returning false when
DIDATA_PASSWORD is a strong password that contains the '%' symbol.
Before this change, getting an entry with that symbol would cause
a ConfigParser.InterpolationSyntaxError which would cause the function
to return false. The Ansible user would then get a confusion "User
credentials not found" error. Since we don't expect the user to employ
any string interpolation in this file, we might as well do away with it.